feat: expose 'run_provider'
This is definitely a hack to get the behavior I want.

A better solution would be to separate the concept of showing text in a
hover window from the concept of getting the text.

Something like:

require('hover.actions').show_text(text_record)

where the text record would have the info currently stored on providers
(bufnr, focusable, etc.)